backend/requirements.txt
# LangGraph Multi-Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Anthropic Claude API
anthropic>=0.40.0

# OpenAI API
openai>=1.50.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
openpyxl>=3.1.0

# API Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=24.1.0

# Utilities
typing-extensions>=4.0.0
pywin32>=306
